Update run.ps1
This came up in an escalation. I modified the if statement 

from:
'if ($item.Name -notin $watchlist.results.SearchKey -and $item.Workload -notin $watchlist.results.Workload)'

to:
'if ($item.Name -notin $watchlist.results.SearchKey)'

The reason for this change was the presence of the workload was preventing new analytics rules based on prior existing workloads to not run due to the watchlist lookup failing. We traced it back to this condition: "-and $item.Workload -notin $watchlist.results.Workload"